Refactor AppRouter and Layout components: streamline routing structure, remove unused imports, and enhance theme handling in Layout.
